+++ dlt/pipeline/mark.py
@@ -1,2 +1,2 @@
-"""Marks data to be specially processed"""
+"""Module with market functions that make data to be specially processed"""
 from dlt.extract.source import with_table_name
