+++ dlt/pipeline/sources.py
@@ -1,0 +1,2 @@
+"""Module with built in sources and source utilities"""
+from dlt.extract.incremental import Incremental as incremental
