+++ tox.ini
@@ -5,4 +5,4 @@ banned-modules = datetime = use dlt.common.pendulum
                  decimal = use dlt.common.decimal
                  decimal.Decimal = use dlt.common.Decimal
                  open = use dlt.common.open
-extend-immutable-calls = dlt.Incremental
+extend-immutable-calls = dlt.sources.incremental
